Remove redundant free operation
According to the C99/C11 Standard (§ 7.20.3 (28)),
The free function causes the space pointed to by ptr to be deallocated,
that is, made available for further allocation. If ptr is a null pointer, no action occurs.

Therefore, if the new element is NULL, the free fucntion in memory allocation for a new element is redundant.